refactor: reorganised content
