--- packages/treeview/src/keyboardInteractions/index.ts
@@ -1,13 +1,0 @@
-/**
- * Copyright Zendesk, Inc.
- *
- * Use of this source code is governed under the Apache License, Version 2.0
- * found at http://www.apache.org/licenses/LICENSE-2.0.
- */
-
-export * from './handleArrowDown';
-export * from './handleArrowLeft';
-export * from './handleArrowRight';
-export * from './handleArrowUp';
-export * from './handleEnd';
-export * from './handleHome';
